+++ script/__init__.py
@@ -1,4 +1,5 @@
 from .theme_config import ThemeConfig
+from .package_config import PackageConfig
 from .theme_builder import compile_theme
 from .installer import install_theme, remove_theme
 from .active_theme import active_theme