+++ changelog.md
@@ -3,6 +3,7 @@
 # v1.1.1
 
 - change osd color
+- add cursor icons
 
 # v1.1.0